Move `Click.Type` to top level class
Both `GuiType` and `ItemType` are top level classes. Why should `Click.Type` be different? `Click.Type.RIGHT` also has too many "."s.
